assets/responsibleai/environments/responsibleai-tabular/tests/requirements.txt
azure-ai-ml==1.25.0
azure-identity==1.20.0
azureml-core>=1.58.0
azureml-mlflow
